+++ src/Soldier.java
@@ -2,7 +2,7 @@
 public class Soldier extends Sprite {
 	public static final String[] name = {"Resources Root/Soldier/g_up", "Resources Root/Soldier/g_dn", "Resources Root/Soldier/g_lt", "Resources Root/Soldier/g_rt"};
 	
-	public Soldier(double x, double y, int action) {
+	public Soldier(int x, int y, int action) {
 		super(x, y, 80, 118, action, name, 9, 5, "gif");
 	}
